+++ internal/provider/provider.go
@@ -70,14 +70,17 @@ func New(version string) func() *schema.Provider {
 
 			DataSourcesMap: map[string]*schema.Resource{
 				"pingone_environment": base.DatasourceEnvironment(),
+
+				"pingone_schema": sso.DatasourceSchema(),
 			},
 
 			ResourcesMap: map[string]*schema.Resource{
 				"pingone_environment": base.ResourceEnvironment(),
 
-				"pingone_group":      sso.ResourceGroup(),
-				"pingone_population": sso.ResourcePopulation(),
-				"pingone_user":       sso.ResourceUser(),
+				"pingone_group":            sso.ResourceGroup(),
+				"pingone_population":       sso.ResourcePopulation(),
+				"pingone_user":             sso.ResourceUser(),
+				"pingone_schema_attribute": sso.ResourceSchemaAttribute(),
 			},
 		}