+++ CHANGELOG.md
@@ -1,3 +1,10 @@
 ## 0.1.0 (Unreleased)
 
-BACKWARDS INCOMPATIBILITIES / NOTES:
+:fire: Initial provider release :fire:
+
+FEATURES:
+
+* **New Resource:** `pingone_environment`
+* **New Resource:** `pingone_group`
+* **New Resource:** `pingone_population`
+* **New Data Source:** `pingone_environment`